+++ packages/uniforms-bootstrap4/src/AutoField.js
@@ -1,5 +1,5 @@
 import BaseField       from 'uniforms/BaseField';
-import invariant       from 'fbjs/lib/invariant';
+import invariant       from 'invariant';
 import {createElement} from 'react';
 
 import NumField    from './NumField';