+++ packages/uniforms/src/SimpleSchema2Bridge.js
@@ -1,5 +1,5 @@
 import cloneDeep from 'lodash/cloneDeep';
-import invariant from 'fbjs/lib/invariant';
+import invariant from 'invariant';
 
 import Bridge         from './Bridge';
 import joinName       from './joinName';