+++ packages/uniforms/src/SimpleSchemaBridge.js
@@ -1,7 +1,7 @@
 /* global Package */
 
 import cloneDeep from 'lodash/cloneDeep';
-import invariant from 'fbjs/lib/invariant';
+import invariant from 'invariant';
 
 import Bridge         from './Bridge';
 import joinName       from './joinName';