+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Next
+- **Changed:** Added necessary `invariant` and `warning` dependencies as React no longer depends on `fbjs`. [\#475](https://github.com/vazco/uniforms/issues/475)
+
 ## [v1.27.0-rc.1](https://github.com/vazco/uniforms/tree/v1.27.0-rc.1) (2018-10-06)
 - **Added:** Support for `@material-ui/core@3.0.0`.
 - **Changed:** Minimum version of `@material-ui/core` is 1.5.1, due to Babel updated.